removed predicate.h and integrated it into ParsecChar.h
